refactor: share the table cell and mark readers instead of keeping two
