+++ frontend/src/apps/slides/utils/tableStructure.test.ts
@@ -12,8 +12,8 @@ const {
 	runTableCommand,
 	distributeColumns,
 	mergeCells,
-	getCells,
 } = await import('./tableStructure')
+const { getCells } = await import('@/apps/slides/stores/tiptapSetup')
 const { getTableSize, getTableWidth, getTableInfo } = await import('./tableWidths')
 const { CellSelection } = await import('prosemirror-tables')
